+++ kurtosis-module/impl/participant_network/el/nethermind/nethermind_launcher.go
@@ -93,6 +93,7 @@ func (launcher *NethermindELClientLauncher) Launch(
 		nodeInfo.Enode,
 		serviceCtx.GetPrivateIPAddress(),
 		rpcPortNum,
+		wsPortNum,
 	)
 
 	return result, nil